+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## unreleased
 
-- Support `DROP PRIMARY KEY` in MySQL
+- Support `ADD PRIMARY KEY` / `DROP PRIMARY KEY` in MySQL
 - Be aware of implicit `NOT NULL` on `PRIMARY KEY`
 
 ## v0.3.1